requirements-dev.txt
# Standard modules
matplotlib>=3.3.4
numpy>=1.23.1
parse>=1.19.0
pygetdp>=1.0.0
swat_em>=0.6.3
gmsh>=4.10.3,<4.14 # block newest version due to issue with .geo export: https://gitlab.onelab.info/gmsh/gmsh/-/issues/3214
GitPython
splines
# to use pyleecan with the recent python versions we have to stick to the development
# state of the package for now.
SciDataTool @ git+https://github.com/eomys/SciDataTool.git
pyleecan @ git+https://github.com/eomys/pyleecan.git@upgrade-python-version

ipykernel # to run tutorial IPython Notebook

# for doc
Sphinx>=7.2.6
sphinx-rtd-theme

# for tests
pytest-check # for integration tests
pytest-dependency
hypothesis>=6.98.17
tox
virtualenv-pyenv

# for build
wheel
setuptools>=42

# for formatting
black
